+++ srcStatic/GetCommandLineArguments.h
@@ -1,0 +1,4 @@
+#include <string>
+#include <vector>
+
+std::vector<std::string> GetCommandLineArguments();
